set up frontend with chakra ui and updated packages

+++ void.products.backend/controllers/productController.js
@@ -43,7 +43,7 @@ export const updateProduct = async (request, response) => {
   if (!mongoose.Types.ObjectId.isValid(id)) {
     return response
       .status(404)
-      .json({ success: false, message: "Product not found!" });
+      .json({ success: false, message: "Invalid Product Id!" });
   }
 
   try {
@@ -62,13 +62,19 @@ export const updateProduct = async (request, response) => {
 export const deleteProduct = async (request, response) => {
   const { id } = request.params;
 
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return response
+      .status(404)
+      .json({ success: false, message: "Invalid Product Id!" });
+  }
+
   try {
     await Product.findByIdAndDelete(id);
     response.status(200).json({ success: true, message: "Product deleted." });
   } catch (error) {
     console.error("Error in deleting product", error.message);
     response
-      .status(404)
-      .json({ success: false, message: "Product not found!" });
+      .status(500)
+      .json({ success: false, message: "Internal Server Error!" });
   }
 }
